\name{DownloadKGML}
\alias{DownloadKGML}
\title{
Download the latest KGML files
}
\description{
Download the latest KGML files from KEGG database if you want the latest KGML files from KEGG database.
}
\details{
Download the latest KGML files from KEGG database before pathway enrichment analysis.
}
\value{
The latest KGML files from KEGG database.
}
\author{
Wei Jiang
}
