\name{PathNetwork}
\alias{PathNetwork}
\title{
Reconstruct pathways to networks
}
\description{
Reconstruct pathways to networks based on KGML files from KEGG database.
}
\usage{
PathNetwork()
}
\details{
Reconstruct pathways to networks based on KGML files from KEGG database.
}
\value{
The relationship of edges in network.
}

\author{
Wei Jiang
}
