\name{getUGraph}
\alias{getUGraph}
\title{
Obtain the graph of pathways
}
\description{
Obtain the graph of pathways.
}
\usage{
getUGraph(graphList, simpleGraph = TRUE)
}
\arguments{
  \item{graphList}{
Get the list.
}
  \item{simpleGraph}{
Convert the network.
}
}
\value{
The graphList relationship.
}
\author{
Wei Jiang
}
