\name{importLatesData}
\alias{importLatesData}
\title{
Import the latest relationship information.
}
\description{
Import the latest relationship information about node,gene and score.
}
\usage{
importLatesData()
}
\details{
Import the latest relationship information about nodes,genes and their scores based on KGML files.
}
\value{
Import the latest relationship information about node,gene and score.
}
\author{
Wei Jiang
}
