\name{NodeGene}
\alias{NodeGene}
\title{
Restract the relationship between nodes and genes.
}
\description{
Restract the relationship between nodes and genes from KGML files.
}
\usage{
NodeGene()
}
\details{
This function must be used behind the function NodeGeneData.
}
\value{
Restract the relationship between nodes and genes in each network based on the information of KGML files.
}
\author{
Wei Jiang
}
