\name{NodeGeneData}
\alias{NodeGeneData}
\title{
Intergate list of node,gene and the score of node.
}
\description{
Intergate list of node,gene and the score of node based on latest KGML files from KEGG database.
}
\usage{
NodeGeneData()
}
\details{
Intergate list of node,gene and the score of node based on latest KGML files from KEGG database.
}
\value{
List contains the relationship of node,gene and the score of node based on latest KGML files.
}
\author{
Wei Jiang
}
