\name{TPEA-package}
\alias{TPEA-package}
\docType{package}
\title{
TPEA: A novel topology-based pathway enrichment analysis approach
}
\description{
This package descirbed a novel topology-based pathway enrichment analysis approach which integrated the topological property of the pathway and the global position of nodes in pathways. There are three functions to finish this process.
}
\details{
The first function is to calculate AUCSC (the area under the cumulative score of node curve) based on interested gene set and the second function is to measure the significance.
The last function is to visualize the pathway in the result based on the genes you input, such as differentially expressed genes.
}
\author{
Wei Jiang
}
